+++ src/config.ts
@@ -1,21 +1,21 @@
 export class ConfigRepository {
     constructor(private workspace: any) {}
 
-    private getWorkspaceConfigure(): any {
-        return this.workspace.getConfiguration('phpunit');
+    private getWorkspaceConfigure(namespace = 'phpunit'): any {
+        return this.workspace.getConfiguration(namespace);
     }
 
-    get(key: string, defaultValue?: any): any {
-        return this.getWorkspaceConfigure().get(key, defaultValue);
+    get(key: string, defaultValue?: any, namespace = 'phpunit'): any {
+        return this.getWorkspaceConfigure(namespace).get(key, defaultValue);
     }
 
-    put(key: string, value: any): ConfigRepository {
-        this.getWorkspaceConfigure().update(key, value);
+    put(key: string, value: any, namespace = 'phpunit'): ConfigRepository {
+        this.getWorkspaceConfigure(namespace).update(key, value);
 
         return this;
     }
 
-    has(key: string): boolean {
-        return this.getWorkspaceConfigure().has(key);
+    has(key: string, namespace = 'phpunit'): boolean {
+        return this.getWorkspaceConfigure(namespace).has(key);
     }
 }