+++ tests/config.test.ts
@@ -1,6 +1,10 @@
 import { ConfigRepository } from '../src/config';
 
 class FakeWorkspaceConfiguration {
+    instance(namespace = 'phpunit') {
+        return this;
+    }
+
     get() {}
 
     update() {}
@@ -11,8 +15,8 @@ class FakeWorkspaceConfiguration {
 class FakeWorkspace {
     constructor(private configuration: FakeWorkspaceConfiguration) {}
 
-    getConfiguration() {
-        return this.configuration;
+    getConfiguration(namespace = 'phpunit') {
+        return this.configuration.instance(namespace);
     }
 }
 
@@ -23,11 +27,13 @@ describe('Config Test', () => {
         const config = new ConfigRepository(workspace);
 
         spyOn(workspace, 'getConfiguration').and.callThrough();
+        spyOn(configuration, 'instance').and.callThrough();
         spyOn(configuration, 'get').and.returnValue('fake');
 
-        expect(config.get('foo', 'bar')).toBe('fake');
+        expect(config.get('foo', 'bar', 'namespace')).toBe('fake');
 
-        expect(workspace.getConfiguration).toHaveBeenCalledWith('phpunit');
+        expect(workspace.getConfiguration).toHaveBeenCalledWith('namespace');
+        expect(configuration.instance).toHaveBeenCalledWith('namespace');
         expect(configuration.get).toHaveBeenCalledWith('foo', 'bar');
     });
 
@@ -37,11 +43,13 @@ describe('Config Test', () => {
         const config = new ConfigRepository(workspace);
 
         spyOn(workspace, 'getConfiguration').and.callThrough();
+        spyOn(configuration, 'instance').and.callThrough();
         spyOn(configuration, 'update').and.callThrough();
 
-        expect(config.put('foo', 'bar')).toBe(config);
+        expect(config.put('foo', 'bar', 'namespace')).toBe(config);
 
-        expect(workspace.getConfiguration).toHaveBeenCalledWith('phpunit');
+        expect(workspace.getConfiguration).toHaveBeenCalledWith('namespace');
+        expect(configuration.instance).toHaveBeenCalledWith('namespace');
         expect(configuration.update).toHaveBeenCalledWith('foo', 'bar');
     });
 
@@ -51,11 +59,13 @@ describe('Config Test', () => {
         const config = new ConfigRepository(workspace);
 
         spyOn(workspace, 'getConfiguration').and.callThrough();
+        spyOn(configuration, 'instance').and.callThrough();
         spyOn(configuration, 'has').and.returnValue(true);
 
-        expect(config.has('foo')).toBe(true);
+        expect(config.has('foo', 'namespace')).toBe(true);
 
-        expect(workspace.getConfiguration).toHaveBeenCalledWith('phpunit');
+        expect(workspace.getConfiguration).toHaveBeenCalledWith('namespace');
+        expect(configuration.instance).toHaveBeenCalledWith('namespace');
         expect(configuration.has).toHaveBeenCalledWith('foo');
     });
 });